requirements.txt
# OpenClaw Python dependencies
# Install: pip3 install --user -r requirements.txt

requests>=2.28.0
pandas>=1.5.0
tradingview-indicators>=0.1.3
python-telegram-bot>=20.0
python-dotenv>=0.20.0
python-binance>=1.0.17
hyperliquid-python-sdk>=0.22.0
